refactor(application): remove unused controller and move logic to BlackJackGame.
