+++ src/main/kotlin/blackjack/Application.kt
@@ -1,7 +1,31 @@
 package blackjack
 
-import blackjack.controller.BlackJackGame
+import blackjack.controller.GameManager
+import blackjack.model.participant.Participants
+import blackjack.model.participant.Player
+import blackjack.model.result.GameResult
+import blackjack.view.InputView
+import blackjack.view.OutputView
+import kotlin.system.exitProcess
 
 fun main() {
-    BlackJackGame.run()
+    runCatching {
+        val names = InputView.readPlayersNames()
+        val bets = InputView.readPlayersBet(names)
+
+        val players =
+            names.zip(bets) { name, amount ->
+                Player(name, amount)
+            }
+
+        val participants = Participants(players)
+
+        GameManager(participants).play()
+
+        val gameResult = GameResult(participants)
+        OutputView.showGameResult(gameResult)
+    }.onFailure { exception ->
+        OutputView.showError(exception.message)
+        exitProcess(1)
+    }
 }
